refactor(47): code refactor
