+++ src/main/java/com/pawelnu/BackendProjectManager/entity/ProjectEntity.java
@@ -1,11 +1,12 @@
 package com.pawelnu.BackendProjectManager.entity;
 
-import com.pawelnu.BackendProjectManager.entity.enums.Status;
+import com.pawelnu.BackendProjectManager.enums.Status;
 import jakarta.persistence.*;
-import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.UUID;
 
 @Entity
 @Table(name = "projects")
