+++ src/main/java/com/pawelnu/BackendProjectManager/mapper/ProjectMapper.java
@@ -3,10 +3,11 @@ package com.pawelnu.BackendProjectManager.mapper;
 import com.pawelnu.BackendProjectManager.dto.project.ProjectCreateRequestDTO;
 import com.pawelnu.BackendProjectManager.dto.project.ProjectDTO;
 import com.pawelnu.BackendProjectManager.entity.ProjectEntity;
-import com.pawelnu.BackendProjectManager.entity.enums.Status;
-import java.util.List;
+import com.pawelnu.BackendProjectManager.enums.Status;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
+
+import java.util.List;
 
 @Mapper(componentModel = "spring")
 public interface ProjectMapper {
