+++ src/main/java/com/pawelnu/BackendProjectManager/repository/project/ProjectRepository.java
@@ -1,10 +1,11 @@
 package com.pawelnu.BackendProjectManager.repository.project;
 
 import com.pawelnu.BackendProjectManager.entity.ProjectEntity;
-import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
 
 @Repository
 public interface ProjectRepository
